Re organized some routes and folder structures
