+++ lib/repository-mappers/figshare-mapper.js
@@ -15,7 +15,7 @@ const jsonLdCrosswalk = {
 function parseAuthors(authors) {
     const jsonldAuthors = [];
     authors.forEach((author) => {
-        jsonldAuthors.push( { 
+        jsonldAuthors.push( {
             'name': author.author_name,
             '@type': 'Person'
         });
@@ -31,7 +31,7 @@ export function getFigshareMetadata(jsonLd) {
             figshareMetadata[jsonLdCrosswalk[key]] = jsonLd[key];
         }
     }
-    
+
     return figshareMetadata;
 }
 
@@ -48,8 +48,8 @@ export function getJsonLd(figshareMetadata) {
         if (figshareAttribute && figshareMetadata.hasOwnProperty(figshareAttribute)) {
             if (figshareMetadata[figshareAttribute]) {
                 jsonLd[key] = figshareMetadata[figshareAttribute];
-            } 
-        }     
+            }
+        }
     }
 
     return jsonLd;
