+++ lib/routes/search.js
@@ -1,10 +1,10 @@
 import express from 'express';
-import { searchFigshare } from '../../repository-clients/figshare-client';
-import { getJsonLd } from '../../repository-mappers/figshare-mapper';
+import { searchFigshare } from '../repository-clients/figshare-client';
+import { getJsonLd } from '../repository-mappers/figshare-mapper';
 const router = express.Router();
 
-const jsonLdAttributes = ['name', 'citation', 'author', 
-        'id', 'keywords', 'email', 'type', 'description', 
+const jsonLdAttributes = ['name', 'citation', 'author',
+        'id', 'keywords', 'email', 'type', 'description',
         'codeRepository', 'license', 'dateCreated'];
 
 function getJsonLdFromQueryParams(queryParams) {
